requirements.txt
langgraph>=0.0.40
langchain>=0.1.16
langchain-core>=0.1.52
langchain-openai>=0.1.6

openai>=1.14.0
typing-extensions>=4.9.0

python-dotenv>=1.0.1
streamlit>=1.32.0